+++ lib/kb_gtdbtk/core/gtdbtk_runner.py
@@ -96,10 +96,12 @@ def _process_output_files(temp_output, out_dir, id_to_name):
             summary_json = '{"data": ' + summary_df.to_json(orient='records') + '}'
             sj = json.loads(summary_json)
             for item in sj['data']:
-                # DEBUG
-                #for key in item.keys():
-                #    print ("KEY: '"+key+"' VAL: '"+str(item[key])+"'")
+                # no blank fields.  messes up datatables in index.html
+                for key in item.keys():
+                    if not item.get(key):
+                        item[key] = '-'  # note: this resets data in sj
 
+                # field 'Name' was changed to 'name'
                 #key = 'Name' if 'Name' in item else 'user_genome'
                 if 'name' in item:
                     key = 'name'
@@ -114,8 +116,7 @@ def _process_output_files(temp_output, out_dir, id_to_name):
 
                 item[key] = id_to_name[this_id]  # note: this resets data in sj
 
-            #print ("OUTFILE: '"+outfile+"'")  # DEBUG
-            #print ("JSON: '"+json.dumps(sj)+"'")  # DEBUG
+            # rewrite with updated vals
             with open(outfile, 'w') as out:
                 out.write(json.dumps(sj))